pyproject.toml
[project]
name = "langchain-search-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "langchain-tavily>=0.2.17",
    "langsmith[claude-agent-sdk]>=0.7.6",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "tavily>=1.1.0",
]
